Include findDdo handler inside handler file.
